backend/requirements.txt
google-generativeai
flask
requests
pygooglenews
python-dotenv
openai